one-step-compile and one-step-secondary

+++ share/mktsimple/one-step-compile.mk
@@ -7,7 +7,7 @@ endif
 
 # Definitions
 SHELL = /bin/bash
-.SHELLFLAGS := -ec
+.SHELLFLAGS ::= -ec
 RMDIR = rm -rf
 MKDIR = mkdir -p
 RM = rm -f
@@ -15,7 +15,8 @@ RM = rm -f
 .SUFFIXES:
 .DELETE_ON_ERROR:
 
-define nl :=
+empty ::=
+define nl ::=
 
 
 endef
@@ -24,7 +25,7 @@ endef
 all:
 
 # Check command-line options
-single_make_options := $(firstword -$(MAKEFLAGS))
+single_make_options ::= $(firstword -$(MAKEFLAGS))
 ifeq (s,$(findstring s,$(single_make_options)))
   silent = 1
 else
@@ -32,7 +33,7 @@ else
 endif
 
 # First info should come before other includes
-makefile_this := $(lastword $(MAKEFILE_LIST))
+makefile_this ::= $(lastword $(MAKEFILE_LIST))
 ifndef silent
   ifndef MAKE_RESTARTS
     $(info **** Starting makefile: '$(makefile_this)' directory '$(CURDIR)')
@@ -42,23 +43,23 @@ ifndef silent
 endif
 
 # setup bindir
-mktsimple_bindir := $(if $(MKTSIMPLE_HOME),$(MKTSIMPLE_HOME)/bin/)
+mktsimple_bindir ::= $(if $(MKTSIMPLE_HOME),$(MKTSIMPLE_HOME)/bin/)
 
-production_goals := all deps
-header_goals := sys-units user-units header-units
-cleanup_goals := clean purge clean-header-units
-action_goals := $(production_goals) $(header_goals) $(cleanup_goals)
-phony_goals = $(action_goals) show
+production_goals ::= all deps
+header_goals ::= sys-units user-units header-units
+cleanup_goals ::= clean purge clean-header-units
+action_goals ::= $(production_goals) $(header_goals) $(cleanup_goals)
+phony_goals ::= $(action_goals) show
 .PHONY: $(phony_goals)
 
 # check goals for cleanup request
-goals := $(MAKECMDGOALS)
+goals ::= $(MAKECMDGOALS)
 ifeq (,$(goals))
-  goals := all
+  goals ::= all
 endif
-cleanup := $(filter $(cleanup_goals),$(goals))
-header_build := $(filter $(header_goals),$(goals))
-not_include_deps := $(or $(cleanup),$(header_build))
+cleanup ::= $(filter $(cleanup_goals),$(goals))
+header_build ::= $(filter $(header_goals),$(goals))
+not_include_deps ::= $(or $(cleanup),$(header_build))
 ifdef cleanup
   ifneq (,$(filter-out $(cleanup_goals),$(goals)))
     $(error ERROR: The goals 'clean', 'purge' and 'clean-header-units' must not be used in conjunction with other targets.)
@@ -71,14 +72,14 @@ ifdef header_build
 endif
 
 # Setup configuration with project file...
-makefile_defs := project.mk
+makefile_defs ::= project.mk
 -include $(makefile_defs)
 
 # Check whether it is a g++ or clang++ compiler
-cxxpath := $(patsubst ./%,%,$(dir $(CXX)))
-cxxfile := $(notdir $(CXX))
-use_clang := $(filter clang%,$(cxxfile))
-use_gcc := $(filter gcc% g++%,$(cxxfile))
+cxxpath ::= $(patsubst ./%,%,$(dir $(CXX)))
+cxxfile ::= $(notdir $(CXX))
+use_clang ::= $(filter clang%,$(cxxfile))
+use_gcc ::= $(filter gcc% g++%,$(cxxfile))
 ifndef cleanup
   ifeq (,$(or $(use_clang),$(use_gcc)))
     $(error CXX must be gcc%/g++% or clang%/clang++% but is $(CXX))
@@ -97,7 +98,7 @@ SRCDIRS ?= .
 # get the last path component from the abspath of the current dir as target name
 TARGET ?= $(lastword $(subst /, ,$(CURDIR)))
 ifdef use_gcc
-  CXX_MODULE_CACHE_DIR := gcm.cache
+  CXX_MODULE_CACHE_DIR ::= gcm.cache
 else ifdef use_clang
   CXX_MODULE_CACHE_DIR ?= pcm.cache
 else
@@ -105,7 +106,7 @@ else
 endif
 
 ifndef SOURCES
-  SOURCES := \
+  SOURCES ::= \
     $(patsubst $(CURDIR)/%,%,\
       $(foreach dir,$(SRCDIRS),\
         $(abspath \
@@ -115,28 +116,27 @@ ifndef SOURCES
      )
 endif
 
-module_db := module-variables.mk
-objects := $(addsuffix .o,$(SOURCES))
-depfiles := $(addsuffix .dep,$(SOURCES))
-mdepfiles := $(depfiles:.dep=.ddi)
+module_db ::= module-variables.mk
+objects ::= $(addsuffix .o,$(SOURCES))
+depfiles ::= $(addsuffix .dep,$(SOURCES))
+mdepfiles ::= $(depfiles:.dep=.ddi)
 
 ifdef use_clang
-  system_header_targets := $(addprefix $(CXX_MODULE_CACHE_DIR)/,$(addsuffix .pcm,$(CXX_SYSTEM_HEADER_UNITS)))
-  user_header_targets := $(addprefix $(CXX_MODULE_CACHE_DIR)/,$(addsuffix .pcm,$(CXX_USER_HEADER_UNITS)))
-  #system_header_file_args := $(foreach head,$(CXX_SYSTEM_HEADER_UNITS),'-fmodule-file=<$(head)>=$(CXX_MODULE_CACHE_DIR)/$(head).pcm')
-  #user_header_file_args := $(foreach head,$(CXX_USER_HEADER_UNITS),-fmodule-file=$(head)=$(CXX_MODULE_CACHE_DIR)/$(head).pcm)
-  #header_file_args := $(addprefix -fmodule-file=,$(system_header_targets)) $(addprefix -fmodule-file=,$(user_header_targets))
-  system_header_file_args := $(addprefix -fmodule-file=,$(system_header_targets))
-  user_header_file_args := $(addprefix -fmodule-file=,$(user_header_targets))
-  header_file_args := $(system_header_file_args) $(user_header_file_args)
+  system_header_targets ::= $(addprefix $(CXX_MODULE_CACHE_DIR)/,$(addsuffix .pcm,$(CXX_SYSTEM_HEADER_UNITS)))
+  user_header_targets ::= $(addprefix $(CXX_MODULE_CACHE_DIR)/,$(addsuffix .pcm,$(CXX_USER_HEADER_UNITS)))
+  system_header_file_args ::= $(addprefix -fmodule-file=,$(system_header_targets))
+  user_header_file_args ::= $(addprefix -fmodule-file=,$(user_header_targets))
+  header_file_args ::= $(system_header_file_args) $(user_header_file_args)
 else
-  system_header_targets := $(addsuffix _target,$(CXX_SYSTEM_HEADER_UNITS))
-  user_header_targets := $(addsuffix _target,$(CXX_USER_HEADER_UNITS))
+  system_header_targets ::= $(addsuffix _target,$(CXX_SYSTEM_HEADER_UNITS))
+  user_header_targets ::= $(addsuffix _target,$(CXX_USER_HEADER_UNITS))
   .PHONY: all sys_units user_units clean purge
 endif
 
 # prevent implicit rules search for makefiles
 .PHONY: $(makefile_defs) $(makefile_this)
+
+CXX_SRC_MOD_CMI_IF_LIST ::=
 
 ifndef not_include_deps
   # Must come before other rules
@@ -145,49 +145,44 @@ ifndef not_include_deps
   include $(depfiles)
 endif
 
-nomodsrcs := $(filter-out $(CXX_MODULE_SOURCES),$(SOURCES))
-nomodobjs :=$(addsuffix .o,$(nomodsrcs))
-modsrcs := $(filter $(CXX_MODULE_SOURCES),$(SOURCES))
-modobjs := $(addsuffix .o,$(modsrcs))
+# Expands to '1' if param1 and param2 are equal or to the empty string otherwise
+# call eq,param1,param2
+eq = $(if $(subst $(2),,$(1)),,1)
+
+# Pretty print the src-mod-cmi-is_if list
+# call pp_src-mod-cmi-is_if list
+pp_src-mod-cmi-is_if = $(foreach line,$1,\
+  $(let src mod cmi is_if,$(subst ;, ,$(line)),\
+    $(info $(empty)	src = '$(src)' mod = '$(mod)' cmi = '$(cmi)' is_if = '$(is_if)')\
+  )\
+)
 
 ifndef silent
   $(info )
-  $(info Build target :$(nl)'$(TARGET)')
-  $(info From sources :$(nl)$(foreach x,$(sort $(SOURCES)),'$(x)'))
-  $(info In directories :$(nl)$(foreach x,$(SRCDIRS),'$(x)'))
-  $(info Module interface units :$(nl)$(foreach x,$(sort $(CXX_MODULE_INTERFACE_UNITS)),'$(x)'))
-  $(info Module units :$(nl)$(foreach x,$(sort $(CXX_MODULE_SOURCES)),'$(x)'))
-  $(info Module names (internal):$(nl)$(foreach x,$(sort $(CXX_MODULES)),'$(x)'))
-  $(info Not a module :$(nl)$(foreach x,$(sort $(nomodsrcs)),'$(x)'))
-  $(info Sytem Header Units :$(nl)$(CXX_SYSTEM_HEADER_UNITS))
-  $(info User Header Units :$(nl)$(CXX_USER_HEADER_UNITS))
+  $(info Build target       : '$(TARGET)')
+  $(info From sources       : $(foreach x,$(sort $(SOURCES)),'$(x)'))
+  $(info In directories     : $(foreach x,$(SRCDIRS),'$(x)'))
+  $(info Sytem Header Units : $(CXX_SYSTEM_HEADER_UNITS))
+  $(info User Header Units  : $(CXX_USER_HEADER_UNITS))
   ifdef verbose
-    $(info Object files:$(nl)$(foreach x,$(sort $(objects)),'$(x)'))
-    $(info Dependency files:$(nl)$(foreach x,$(sort $(depfiles)),'$(x)'))
-    $(info Module dependency files:$(nl)$(foreach x,$(sort $(mdepfiles)),'$(x)'))
-    $(info Compiled module interface files:$(nl)$(foreach x,$(sort $(CXX_CMI_FILES)),'$(x)'))
-    $(info This makefile:$(nl)'$(makefile_this)')
-    $(info Sytem Header Targets:$(nl)$(system_header_targets))
-    $(info User Header Targets:$(nl)$(user_header_targets))
-    $(info )
-    $(info Sourcefile to modulname database (CXX_SOURCE2MODULE_<source name>))
-    $(foreach src,$(CXX_MODULE_SOURCES),$(eval $(info $(src) -> $(CXX_SOURCE2MODULE_$(src)))))
-    $(info Sourcefile to CMI-file database (CXX_SOURCE2CMI_<source name>))
-    $(foreach src,$(CXX_MODULE_SOURCES),$(eval $(info $(src) -> $(CXX_SOURCE2CMI_$(src)))))
-    $(info Modulname to CMI-file database (CXX_MODULE2CMI_<modulname>))
-    $(foreach mod,$(CXX_MODULES),$(eval $(info $(mod) -> $(CXX_MODULE2CMI_$(mod)))))
-    $(info MAKE_TERMOUT: $(MAKE_TERMOUT) MAKE_TERMERR: $(MAKE_TERMERR))
-    $(info MAKE_VERSION: $(MAKE_VERSION))
+    $(info Object files            : $(foreach x,$(sort $(objects)),'$(x)'))
+    $(info Dependency files        : $(foreach x,$(sort $(depfiles)),'$(x)'))
+    $(info Module dependency files : $(foreach x,$(sort $(mdepfiles)),'$(x)'))
+    $(info This makefile        : '$(makefile_this)')
+    $(info Sytem Header Targets : $(system_header_targets))
+    $(info User Header Targets  : $(user_header_targets))
+    $(info CXX_SRC_MOD_CMI_IF_LIST:)
+    $(call pp_src-mod-cmi-is_if,$(CXX_SRC_MOD_CMI_IF_LIST))
+    $(info MAKE_TERMOUT : $(MAKE_TERMOUT) MAKE_TERMERR : $(MAKE_TERMERR))
+    $(info MAKE_VERSION : $(MAKE_VERSION))
   endif
   $(info )
 endif
 
-# db check here
-
 # set c++ standard to c++20 if not done otherwise
-cppstd :=
+cppstd ::=
 ifeq (,$(filter -std=%,$(CXXFLAGS)))
-  cppstd := -std=c++20 # One space is appended!
+  cppstd ::= -std=c++20 # One space is appended!
 endif
 
 # Rules section
@@ -197,12 +192,12 @@ deps: $(depfiles)
 
 # Dep files depend on the source and the generated rules with -MQ $*
 ifdef use_clang
-  scandeps := $(cxxpath)clang-scan-deps$(patsubst clang++%,%,$(cxxfile))
+  scandeps ::= $(cxxpath)clang-scan-deps$(patsubst clang++%,%,$(cxxfile))
   dep_recipe = $(scandeps) -o $*.ddi -format=p1689 -- $(CXX) -o $*.o $< -MMD -MF $*.dep -MQ $*.dep -MQ $*.ddi -MP -c $(cppstd)$(CXXFLAGS) $(CPPFLAGS) $(TARGET_ARCH) $(header_file_args)
-  cmi_extension := pcm
+  cmi_extension ::= pcm
 else
   dep_recipe = $(CXX) $< -MM -MF '$*.dep' -MQ $*.dep -MQ $*.ddi -MP -fdeps-format=p1689r5 -fdeps-file=$*.ddi -fdeps-target=$*.o -c -fmodules $(cppstd)$(CXXFLAGS) $(CPPFLAGS) $(TARGET_ARCH)
-  cmi_extension := gcm
+  cmi_extension ::= gcm
 endif
 
 %.dep %.ddi: %
@@ -217,6 +212,69 @@ $(module_db): $(mdepfiles)
 	$(mktsimple_bindir)p1689_to_make_db.sh $(cmi_extension) $(CXX_MODULE_CACHE_DIR) $@ $^
 	$(if $(silent),,@echo -e "Finished database preparation in $@\n")
 
+# Rules with Grouped Targets for modules to translate a module source into object and CMI file
+ifdef use_clang
+# call source-file,cmi-file
+define modul_template =
+$(src).o $(cmi) &: $(src) $(src).dep | $(CXX_MODULE_CACHE_DIR)
+	$$(if $$(silent),,@$$(RM) $$(verbose) $(src).o $(cmi))
+	$$(CXX) -o $(src).o -fmodule-output='$(cmi)' -x c++-module $(src) -c -fprebuilt-module-path=$$(CXX_MODULE_CACHE_DIR) -fmodules-reduced-bmi $$(cppstd)$$(CXXFLAGS) $$(CPPFLAGS) $$(TARGET_ARCH) $$(header_file_args)
+	$$(if $$(silent),,@echo -e "Finished modul compiling: $(src)\n")
+endef
+else
+define modul_template =
+$(src).o $(cmi) &: $(src) $(src).dep | $(CXX_MODULE_CACHE_DIR)
+	$$(if $$(silent),,@$$(RM) $$(verbose) $(src).o $(cmi))
+	$$(CXX) -o $(src).o $(src) -c -fmodules $$(cppstd)$$(CXXFLAGS) $$(CPPFLAGS) $$(TARGET_ARCH)
+	$$(if $$(silent),,@echo -e "Finished modul compiling: $(src)\n")
+endef
+endif
+
+modules ::=
+modsrcs ::=
+mod_if_units ::=
+
+define database_template
+  modules += $(mod)
+  modsrcs += $(src)
+  $(if $(call eq,$(is_if),1),mod_if_units += $(src))
+endef
+
+# Generate rules for all modules
+# call src_mod_cmi_if_list
+make_module_rule = $(foreach line,$1,\
+  $(let src mod cmi is_if,$(subst ;, ,$(line)),\
+    $(if $(silent),,$(info Generate Module Rule $(src).o $(cmi) &: $(src) $(src).dep))\
+    $(if $(and $(src),$(mod),$(cmi),$(is_if)),\
+      ,\
+      $(error Inconsistent CXX_SRC_MOD_CMI_IF_LIST : '$(src)' '$(mod)' '$(cmi)' '$(is_if)')\
+    )\
+    $(eval $(modul_template))\
+    $(eval $(database_template))\
+  )\
+)
+
+# Escape module db for eval
+escaped_src_mod_cmi_if_list ::= $(subst $$,$$$$,$(CXX_SRC_MOD_CMI_IF_LIST))
+$(call make_module_rule,$(escaped_src_mod_cmi_if_list))
+
+nomodsrcs ::= $(filter-out $(modsrcs),$(SOURCES))
+nomodobjs ::=$(addsuffix .o,$(nomodsrcs))
+modobjs ::= $(addsuffix .o,$(modsrcs))
+
+ifndef silent
+  $(info )
+  $(info Module units            : $(foreach x,$(sort $(modsrcs)),'$(x)'))
+  $(info Module interface units  : $(foreach x,$(sort $(mod_if_units)),'$(x)'))
+  $(info Module names (internal) : $(foreach x,$(sort $(modules)),'$(x)'))
+  $(info Not a module            : $(foreach x,$(sort $(nomodsrcs)),'$(x)'))
+  ifdef verbose
+    $(info Modulname to CMI-file database (CXX_MODULE2CMI_<modulname>))
+    $(foreach mod,$(modules), $(info $(empty)	$(mod) -> $(CXX_MODULE2CMI_$(mod))))
+  endif
+  $(info )
+endif
+
 # generate objects from non module sources
 ifdef use_clang
   nomodobjs_recipe = $(CXX) $(OUTPUT_OPTION) $< -c -fprebuilt-module-path=$(CXX_MODULE_CACHE_DIR) -fmodules-reduced-bmi $(cppstd)$(CXXFLAGS) $(CPPFLAGS) $(TARGET_ARCH) $(header_file_args)
@@ -227,27 +285,6 @@ $(nomodobjs): %.o: % %.dep
 	@$(RM) $(verbose) $@
 	$(nomodobjs_recipe)
 	$(if $(silent),,@echo -e "Finished building: $<\n")
-	
-
-# Rules with Grouped Targets for modules to translate a module source into object and CMI file
-ifdef use_clang
-# call source-file,cmi-file
-define modul_template =
-$(1).o $(2) &: $(1) $(1).dep | $(CXX_MODULE_CACHE_DIR)
-	$$(if $$(silent),,@$$(RM) $$(verbose) $(1).o $(2))
-	$$(CXX) -o $(1).o -fmodule-output='$(2)' -x c++-module $(1) -c -fprebuilt-module-path=$$(CXX_MODULE_CACHE_DIR) -fmodules-reduced-bmi $$(cppstd)$$(CXXFLAGS) $$(CPPFLAGS) $$(TARGET_ARCH) $$(header_file_args)
-	$$(if $$(silent),,@echo -e "Finished modul compiling: $(1)\n")
-endef
-else
-define modul_template =
-$(1).o $(2) &: $(1) $(1).dep | $(CXX_MODULE_CACHE_DIR)
-	$$(if $$(silent),,@$$(RM) $$(verbose) $(1).o $(2))
-	$$(CXX) -o $(1).o $(1) -c -fmodules $$(cppstd)$$(CXXFLAGS) $$(CPPFLAGS) $$(TARGET_ARCH)
-	$$(if $$(silent),,@echo -e "Finished modul compiling: $(1)\n")
-endef
-endif
-# Generate rules for all modules
-$(foreach src,$(modsrcs),$(eval $(call modul_template,$(src),$$(CXX_SOURCE2CMI_$(src)))))
 
 $(TARGET): $(modobjs) $(nomodobjs)
 	@$(RM) $(verbose) $@
@@ -286,26 +323,30 @@ endef
 endif
 
 $(CXX_MODULE_CACHE_DIR) :
-	@$(MKDIR) $(verbose) $@
+	$(MKDIR) $(verbose) $@
 	$(if $(silent),,@echo)
 
 clean-header-units:
 	$(if $(silent),,@echo "Header Unit Cleanup")
+ifdef use_clang
 	$(RM) $(verbose) $(system_header_targets)
 	$(RM) $(verbose) $(user_header_targets)
+else
+	for x in $(CXX_MODULE_CACHE_DIR)/*; do if [[ -d $${x} ]]; then $(RMDIR) $(verbose) "$${x}"; fi; done
+endif
 	$(if $(silent),,@echo)
 
 clean:
 	$(if $(silent),,@echo "Cleanup")
-	$(RM) $(verbose) $(TARGET) $(CXX_CMI_FILES)
-	$(RM) $(verbose) $(nomodobjs) $(modobjs)
+	$(RM) $(verbose) $(CXX_MODULE_CACHE_DIR)/*.$(cmi_extension)
+	$(RM) $(verbose) $(objects)
+	$(RM) $(verbose) $(depfiles)
+	$(RM) $(verbose) $(p1689files)
+	$(RM) $(verbose) $(module_db)
 	$(if $(silent),,@echo)
 
 purge: clean clean-header-units
 	$(if $(silent),,@echo "Purge")
-	$(RM) $(verbose) $(depfiles)
-	$(RM) $(verbose) $(mdepfiles)
-	$(RM) $(verbose) $(module_db)
 	$(RM) $(verbose) *.o *.dep *.ddi
 	$(RMDIR) $(verbose) $(CXX_MODULE_CACHE_DIR)
 	$(if $(silent),,@echo)
